코드 정리 및 KIS_API 기능 목록 생성
